requirements.txt
fastapi>=0.100.0
pydantic>=2.4.2,<3.0.0
crewai
uvicorn[standard]==0.22.0
sqlalchemy==2.0.27
pymysql==1.1.0
python-dotenv==1.0.0
passlib==1.7.4
python-jose==3.3.0
email-validator==2.1.0
typing-extensions>=4.6.0
pyasn1
ecdsa!=0.15
PyJWT==2.4.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
requests>=2.31.0